Fixed getDamager(0 undefined.
This whole PR was a mess. I just noticed getDamager() is different to getModifier(). I've tested similar plugins that has this part of the code and they seem to work just fine. I'm not entirely sure what happened, but this PR should fix everything relating to the issue posts you were receiving.
Let's just see how this plays out, and let me know if this fixes it or not. Thanks.